refactor: [SG-25367] Reorganize utility modules into utils/ package
Move publish utility functions into a dedicated utils/ package alongside
the new tga_image module, replace temp-file generation in TGA tests with
checked-in binary fixtures, and rename tga_to_pixmap to tga_to_qpixmap.

Changes:
- Move util.py to utils/publish.py
- Add alias `from .utils import publish as util` in package __init__ to
  preserve backward compatibility for app.util.* and publisher.util.*
  public API used by hooks and external configs
- Expose publish and tga_image submodules in utils/__init__.py to match
  the existing api/publish_tree_widget attribute-access pattern
- Update docs/utility.rst automodule target to new module path
- Rename tga_image.tga_to_pixmap -> tga_to_qpixmap for clarity
- Replace _make_tga_file struct.pack helper with two committed TGA
  fixtures (test_rle.tga, test_uncompressed.tga) to avoid the circular
  test dependency where the encoder used the same format knowledge as
  the decoder under test
