+++ python/tk_multi_publish2/__init__.py
@@ -12,7 +12,7 @@ import sgtk
 
 from .api import PublishManager  # noqa
 from . import base_hooks  # noqa
-from . import util  # noqa
+from .utils import publish as util  # noqa
 from . import publish_tree_widget  # noqa
