+++ python/tk_multi_publish2/utils/__init__.py
@@ -8,4 +8,5 @@
 # agreement to the Shotgun Pipeline Toolkit Source Code License. All rights
 # not expressly granted therein are reserved by Shotgun Software Inc.
 
+from . import publish  # noqa
 from . import tga_image  # noqa
