+++ tests/test_item.py
@@ -455,7 +455,7 @@ class TestQtPixmapAvailability(PublishApiTestBase):
         fake_pixmap = self.QtGui.QPixmap(self.image_path)
         tga_image = self.app.import_module("tk_multi_publish2").utils.tga_image
         with patch.object(tga_image, "is_tga_rle", return_value=True), patch.object(
-            tga_image, "tga_to_pixmap", return_value=fake_pixmap
+            tga_image, "tga_to_qpixmap", return_value=fake_pixmap
         ):
             item = self.PublishItem("test", "test", "test")
             item.set_thumbnail_from_path(__file__)
